refactor(sign) : 코드 다듬기
